+++ packages/python/src/sandkiln/__init__.py
@@ -1,6 +1,7 @@
 from .drive import Drive, DriveHolder, DriveInfo
 from .errors import SandkilnApiError
 from .image import Image, ImageInfo
+from .pool import Pool, PoolInfo
 from .sandbox import DirEntry, DriveAttachment, ExecResult, Sandbox, SandboxInfo, SnapshotInfo, StopResult
 
 __all__ = [
@@ -16,5 +17,7 @@ __all__ = [
     "Drive",
     "DriveInfo",
     "DriveHolder",
+    "Pool",
+    "PoolInfo",
     "SandkilnApiError",
 ]